Combine logical OR ranges properly. pr97567
update testcase to work on 32 bit targets

	gcc/testsuite
	* gcc.dg/pr97567.c: Update to work with 32 bit targets.

+++ gcc/testsuite/gcc.dg/pr97567.c
@@ -4,7 +4,7 @@
 int a, b, c, d;
 void k() {
   unsigned f = 1;
-  long g = 4073709551615;
+  long long g = 4073709551615;
   for (; a; a++)
     for (;;) {
       d = 0;
@@ -16,7 +16,7 @@ void k() {
       ;
   g || f;
   int i = 0 - f || g;
-  long j = g - f;
+  long long j = g - f;
   if (j || f) {
     if (g < 4073709551615)
       for (;;)
